+++ NEWS.md
@@ -9,6 +9,7 @@ border-radius: 128px;
 ## v1.10.9
 
 - Fixes
+    - Fix browser settings not clearing with clear button.
     - Fix search settings not clearing on OPDS start page.
 
 ## v1.10.8